Add KK Party compatibility (#56)

+++ Shared/Interpreters/HSceneInterpreter.cs
@@ -421,7 +421,14 @@ namespace KK_VR.Interpreters
         
         internal static void OnPoseChange(HSceneProc.AnimationListInfo anim)
         {
-            mode = anim.mode switch
+            //mode = anim.mode switch
+            //{
+            //    EMode.houshi or EMode.houshi3P or EMode.houshi3PMMF => EMode.houshi,
+            //    EMode.sonyu or EMode.sonyu3P or EMode.sonyu3PMMF => EMode.sonyu,
+            //    _ => anim.mode,
+            //};
+
+            mode = (int)anim.mode switch
             {
                 EMode.houshi or EMode.houshi3P or EMode.houshi3PMMF => EMode.houshi,
                 EMode.sonyu or EMode.sonyu3P or EMode.sonyu3PMMF => EMode.sonyu,
